refactor: move parameter classes to exports
